Finalisation of the task

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -2,46 +2,46 @@ package ru.netology.stats;
 
 public class StatsService {
 
-    public int getAmountSales(long[] sales) {        //Подсчет суммы продаж
+    public long getAmountSales(long[] sales) {        //Подсчет суммы продаж
 
-        int amount = 0;
+        long amount = 0;
         for (int i = 0; i < sales.length; i++) {
             amount = (int) (amount + sales[i]);
         }
         return amount;
     }
 
-    public int getAverageSales(long[] sales) {       //Подсчет средней суммы продаж
-        int average = getAmountSales(sales) / 12;
+    public long getAverageSales(long[] sales) {       //Подсчет средней суммы продаж
+        long average = getAmountSales(sales) / 12;
 
         return average;
     }
 
-    public int getMaxSales(long[] sales) {          //Подсчет номера месяца, в котором был пик продаж
-        int maxMonth = 0;
+    public long getMaxSales(long[] sales) {          //Подсчет номера месяца, в котором был пик продаж
+        long maxMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >= sales[maxMonth]) {
+            if (sales[i] >= sales[(int) maxMonth]) {
                 maxMonth = i;
             }
         }
         return maxMonth + 1;
     }
 
-    public int getMinSales(long[] sales) {         //Подсчет номера месяца, в котором был минимум продаж
-        int minMonth = 0;
+    public long getMinSales(long[] sales) {         //Подсчет номера месяца, в котором был минимум продаж
+        long minMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= sales[minMonth]) {
+            if (sales[i] <= sales[(int) minMonth]) {
                 minMonth = i;
             }
         }
         return minMonth + 1;
     }
 
-    public int getMinAverage(long[] sales) {      //Количество месяцев, в которых продажи были ниже среднего
-        int countMinAverage = 0;
-        int average = getAverageSales(sales);
+    public long getMinAverage(long[] sales) {      //Количество месяцев, в которых продажи были ниже среднего
+        long countMinAverage = 0;
+        long average = getAverageSales(sales);
         for (int i = 0; i < 12; i++) {
             if (sales[i] < average) {
                 countMinAverage++;
@@ -50,9 +50,9 @@ public class StatsService {
         return countMinAverage;
     }
 
-    public int getMaxAverage(long[] sales) {      //Количество месяцев, в которых продажи были выше среднего
-        int countMaxAverage = 0;
-        int average = getAverageSales(sales);
+    public long getMaxAverage(long[] sales) {      //Количество месяцев, в которых продажи были выше среднего
+        long countMaxAverage = 0;
+        long average = getAverageSales(sales);
         for (int i = 0; i < 12; i++) {
             if (sales[i] >= average) {
                 countMaxAverage++;
